+++ MyCraft/src/Player/Colisions.hpp
@@ -5,6 +5,6 @@
 
 class colisions {
     public:
-    bool isSolidAround(glm::vec3 pos, const std::map<std::pair<int, int>, Chunk>& World, glm::vec3 ChunkSize, float margin = 0.25f, float height = 1.8f);
-    bool isSolidAt(glm::vec3 pos, const std::map<std::pair<int, int>, Chunk>& World, glm::vec3 ChunkSize);
+    bool isSolidAround(glm::vec3 pos, const std::map<std::pair<int, int>, Chunk>& World, glm::ivec3 ChunkSize, float margin = 0.25f, float height = 1.8f);
+    bool isSolidAt(glm::vec3 pos, const std::map<std::pair<int, int>, Chunk>& World, glm::ivec3 ChunkSize);
 };
